requirements.txt
attrs
pytest